Реализована проверка добавления и обновления продуктов и товаров в админке

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Category;
+use App\Http\Requests\CategoryRequest;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Storage;
@@ -36,11 +37,17 @@ class CategoryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CategoryRequest $request)
     {
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+
+        unset($params['image']);
+
+        if ($request->has('image')) {
+            $path = $request->file('image')->store('categories');
+            $params['image'] = $path;
+        }
+
         Category::create($params);
         return redirect()->route('categories.index');
     }
@@ -74,12 +81,18 @@ class CategoryController extends Controller
      * @param  \App\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category)
+    public function update(CategoryRequest $request, Category $category)
     {
-        Storage::delete($category->image);
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+
+        unset($params['image']);
+
+        if ($request->has('image')) {
+            Storage::delete($category->image);
+            $path = $request->file('image')->store('categories');
+            $params['image'] = $path;
+        }
+
         $category->update($params);
         return redirect()->route('categories.index');
     }
